requirements.txt
attrs==22.1.0
contourpy==1.0.6
cycler==0.11.0
exceptiongroup==1.0.1
fonttools==4.38.0
iniconfig==1.1.1
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
tomli==2.0.1
